requirements.txt
alembic==1.5.5
astroid==2.5.1
Babel==2.9.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-User==1.0.2.2
Flask-WTF==0.14.3
idna==3.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.4
pathlib==1.0.1
psycopg2==2.8.6
pycparser==2.20
pylint==2.7.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.23
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
